Allow actions to load their module implementations on-demand.
By only registering the action, we skip the processing that would been necessary to load and evaluate the module scripts. This will allow us to improve overall performance while also scaling up the number of supported actions. To force loading of the entire module (to override specific calls, for instance), call require().