Fix unit test to match current behavior
I *think* this unit test has fallen out of date with the latest code. debugdir() is not set for the test, so the working directory should be the same as the project directory, so the attribute should be set to an empty string.

I need this test to pass before I can update remake-core master to use the latest version of this module.