Fix default entry point for Windows console applications
By default, projects created for console applications through Visual
Studio don't explicitly specify the entry point symbol; the linker
automatically detects whether "mainCRTStartup" or "wmainCRTStartup"
should be used based on whether main() or the Windows-specific wmain()
is defined. The current behavior of Premake is to explicitly use
"mainCRTStartup" in both windowed and console applications if the
"WinMain" flag is not set, preventing the use of wmain() in console
applications unless the entry point is explicitly set in the project
configuration.

This change makes it so Premake does not explicitly specify the entry
point symbol for Windows console applications, allowing wmain() to be
used without the need for additional configuration settings.