Support for username/password in HTTP requests
http.get and http.download now accept an optional additional parameter which allows specifying the user name and password that will used in the request in the format of username:password.

This utilizes the default authentication of curl which is basic authentication.
This functionality will be utilized for authenticated http retrieval from GitHub APIs.