Fixes for authentication functionality.
The userpwd paramater for http.get and http.download is now correctly optional.

Verified that it doesn't collide with the optional headers parameter to these functions because the type signature is different (headers is passed in as a table).

Added a user agent which is required to interface with GitHub APIs.