Cleanup of http.get and http.download
All options are now provided via a table which currently supports progress, userpwd, username, password, headers and should be extensible to add more options.

The previous signature of these funtions is supported for backward compatibility.

It appears that there were logic errors in the existing code such that a progress function would have only worked as the last parameter and would not work if headers were supplied. This is fixed such that the old signature will work if progress and headers are supplied.