Improvements to http.get and http.download
There was an undocumented return code from http.download which was the
numerical CURL error code.

This has been replaced with the HTTP response code. The return for
http.download is now error message, HTTP response code. This could
break code that was relying on the CURL error but my light research is
that users of this call were checking for the error message being “OK”
and that functionality is preserved.

http.get has been updated to have a consistent return signature to
http.download and now returns the result, error message, HTTP response
code. The error message is now set to “OK” to match http.download.

These changes abstract this api from curl and make the focus http.

Cleaned up the code to make it easier to follow and removed some memory
leaks.

http.get and http.download would leak if headers were supplied.

http.get would leak the string buffer if curl_init failed.