Add NoLinkObjects flag
The default behaviour is to link .obj files if a custom build command
outputs them, but there are some cases where we don't want this to
happen, such as when dealing with Wavefront .obj model files.

This commit will prevent the automatic linking of .obj files that have
this new flag.

When an obj file is outputted by a custom build command, this prevents
the linker from linking it. This is necessary, for example, when
dealing with obj model files.