Fix VS2015 sln icon
This is a minor fix to make the correct sln icon appear in Windows. The
Icon Handler for sln files is apparently quite picky.

VS2015 sln files contain this:
# Visual Studio 14

Not this:
# Visual Studio 2015

Having 2015 instead of 14 does not appear to cause any major issues. But
you get the generic sln file icon in Explorer instead of the
VS2015-specific one.