MSVC generation: Avoid "AdditionalOptions"
Currently the code puts flags into "AdditionalOptions" rather
than outputing attributes like "RuntimeLibrary", "Optimization",
"DebugInformationFormat" and "ExceptionHandling".

This means we get a clash of flags when compiling, and the MSVC
properties pages show the wrong thing.

Here we look at the flags and output the correct things based
on them.

I apologise for my crap lua here. There is probably a better,
cleaner way to do this, but this is the first bit of lua I've
ever done.