Fix VS2005 output of quoted strings.
We have a project file that needs to set a symbol to include quoted
strings. For instance:

   defines {
     'FT_CONFIG_OPTIONS_H="slimftoptions.h"'
   }

The source .c file then does:

   #include FT_CONFIG_OPTIONS_H

With the definition like this, gnu make can accept the generated
files and does the wrong thing. VS project files have the quotes
appear in them as &quot; but these disappear on being passed to
the actual compiler.

Having modified premake5 to output \quot; we get exactly the
results we need.

An alternative workaround we tried was to use:

   defines {
     'FT_CONFIG_OPTIONS_H=\\"slimftoptions.h\\"'
   }

which produced working VS2005 files, but broken gnu make files.