On OSX, convert systemversion to MACOSX_DEPLOYMENT_TARGET
Per the discussion in issue #1336, if systemversion is set
output MACOSX_DEPLOYMENT_TARGET into the resulting xcode project.
This borrows a diff from the (unmerged) PR #1003, applies it
in isolation, and adds a pair of unit tests to confirm that both
setting and not setting the systemversion results in correct
output.

Closes #1336