Allow .dylib extension in links for Xcode
When entries in links are specified with an extension we treat them
like frameworks for the most part. They are placed into the
Frameworks group and added to the Frameworks build phase. We do have
separate logic for the file reference, only because we never want to
resolve a .dylib link to the /System/Frameworks directory. Instead
we assume a .dylib link that is just a file name, e.g. 'libA.dylib',
is relative to the project the same way as if it were prefixed with
'./'.