Fixes issues in VS Script when comparing paths for PCH Source
The source in prjcfg is the same string the user set this means it can
include duplicate / or \, can also have both back/forward-slashes. In
summary, the path is not normalized. Which will cause the matching to
fail and the PCH Source will not be set.

There is also another issue, path.normalize does not remove the double
\ from the drive name. For example, if the path is:

C:\\MySource\\sources.cpp

Normalize will transform it to

C:\\MySources/sources.cpp

ignoring the drive \\, this is why there is a string replace after
calling normalize.