gmake2: Fix detecting msdos vs posix shell
The current method of searching for ".exe" in the ComSpec environment
variable is not reliable, because no shell on Windows I could find
updates this. In theory, even if shells did, this still wouldn't work
because `sh.exe` is a valid name for shell running build rules.

I could not find any posix type shell that actually changes the ComSpec environment
variable on Windows.

All of these yielded the same result:

powershell: $env:ComSpec
busybox-w32: echo $COMSPEC
msys2: echo $COMSPEC
cygwin: echo $COMSPEC

returning: C:\Windows\system32\cmd.exe

Not even powershell touches this environment variable at all!

The only reason this ever appeared to work in the first place, is that
most posix type shells on Windows capitalize environment variables, and
GNU Make does case-sensitive comparisons of environment variables, even
on Windows.  What it effectively was doing is detecting the shell type
of the shell that was invoking make, not actually what was executing the
build rules.

The shell executing the build rules and the shell make is invoked from
don't have to be the same. In fact, on Windows, GNU Make searches the
PATH for an `sh.exe` to run build rules, and prioritizes using it over
cmd.exe if it is present (see variable.c for how this is implemented).
In an environment like this, invoking `make clean` from cmd.exe will
fail, due to trying to execute cmd commands on a posix type shell, which
cannot handle them. This can be proven to be the case, because `make
SHELL=cmd` works correctly, by forcing GNU Make to invoke cmd.exe
instead of its preferred sh.exe

As far as I can tell, there is no standard environment variable that
indicates whether an msdos or posix type shell is being used to execute
the Makefile. The solution I implemented is to use a "polyglot test"
executed in the shell, that can differentiate between an msdos and type
posix shell in the actual Makefile itself.

I used the approach of running `echo "test"` in the shell. A posix type
shell will write test without quotes, while an msdos type shell will
write "test" with quotes.