Fix gmake action mixing up compilers from different toolsets.
Due to Make behavior, specifically its default values to the CC and CXX
implicit rules, we could end up in a scenario where the gcc and clang
compilers would be used in the same project.

Be more explicit and do not rely on using the system default compilers,
and use the gcc and g++ compilers explicitly.