Extracts target dir compute from `config.buildtargetinfo` into `config.targetdir`.
This adds a new internal helper API to get the target dir of a config or
project: `config.targetdir`, which is a refactor of the existing logic.

I needed a way to access the `targetdir` of a project pre-bake, which
was not possible before. With the new `config.targetdir` function its
straightforward to get it.