Adds a new toolset `driver` flag to signal driver compatibility.
This allows external consumers of those toolsets to know what kind of
driver the tool uses.

This can be important in some cases, like in Ninja, which provides some
dependency handling features that need to change depending if the
compiler is GCC or MSVC-like.

The problem is that due to this information, the Ninja generator ends up
having hardcoded logic for this depending on the toolset,
https://github.com/jimon/premake-ninja/blob/master/ninja.lua#L401.

But this is not extensible to other modules that define their own
toolset (like the Emscripten module).