Increase Lua's debug path size.
When Lua reports a stack traceback, it prints the path to the files.

But by default, right now, it's limited to 60 characters, which gets
super annoying when trying to work on the codebase, as paths are
ellipsized if they are bigger, which prevents from Ctrl-clicking on them
to go to the right location in the IDE/editor.

This increases the limit to double the default size, which so far has
worked fine for me.

Before:
```
stack traceback:
        .../premake/modules/gmake/tests/cpp/test_make_pch.lua:66:
```

After:
```
stack traceback:
/home/joao/dev/tools/premake/modules/gmake/tests/cpp/test_make_pch.lua:66:
```